+++ src/const.js
@@ -1,0 +1,3 @@
+const TYPES_EVENT = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
+
+export {TYPES_EVENT};
