+++ src/main.js
@@ -1,11 +1,11 @@
 import TripPresenter from './presenter/trip-presenter';
 import { render } from './render';
-import TripFiltersView from './view/trip-filters-view';
+import TripFiltersView from './view/trip-filters/trip-filters-view';
 
-const filters = document.querySelector('.trip-controls__filters');
-const tripEvents = document.querySelector('.trip-events');
-const tripPresenter = new TripPresenter({ tripContainer: tripEvents });
+const filtersElement = document.querySelector('.trip-controls__filters');
+const tripEventsElement = document.querySelector('.trip-events');
+const tripPresenter = new TripPresenter({ tripContainerElement: tripEventsElement });
 
-render(new TripFiltersView(), filters);
-tripPresenter.init();
+render(new TripFiltersView(), filtersElement);
+tripPresenter.render();
