+++ src/presenter/trip-presenter.js
@@ -1,20 +1,20 @@
 import { render } from '../render';
-import EventCreateView from '../view/event-create-view';
-import EventEditView from '../view/event-edit-view';
-import EventListView from '../view/event-list-view';
-import EventView from '../view/event-view';
-import TripSortView from '../view/trip-sort-view';
+import EventCreateView from '../view/event-create/event-create-view';
+import EventEditView from '../view/event-edit/event-edit-view';
+import EventListView from '../view/event-list/event-list-view';
+import EventView from '../view/event/event-view';
+import TripSortView from '../view/trip-sort/trip-sort-view';
 
 export default class TripPresenter {
   eventListComponent = new EventListView();
 
-  constructor({tripContainer}) {
-    this.tripContainer = tripContainer;
+  constructor({tripContainerElement}) {
+    this.tripContainerElement = tripContainerElement;
   }
 
-  init() {
-    render(new TripSortView(), this.tripContainer);
-    render(this.eventListComponent, this.tripContainer);
+  render() {
+    render(new TripSortView(), this.tripContainerElement);
+    render(this.eventListComponent, this.tripContainerElement);
     render(new EventEditView(), this.eventListComponent.getElement());
     render(new EventCreateView(), this.eventListComponent.getElement());
