+++ src/utils.js
@@ -1,0 +1,7 @@
+function capitalize(str){
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
+export {capitalize};
+
+
