+++ src/view/event-create/event-create-view.js
@@ -1,0 +1,20 @@
+import { createElement } from '../../render';
+import { createEventCreateTemplate } from './event-create-template';
+
+export default class EventCreateView {
+  getTemplate() {
+    return createEventCreateTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
