+++ src/view/event-edit/event-edit-view.js
@@ -1,0 +1,20 @@
+import { createElement } from '../../render';
+import { createEventEditTemplate } from './event-edit-template';
+
+export default class EventEditView {
+  getTemplate() {
+    return createEventEditTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
