+++ src/view/event-list/event-list-template.js
@@ -1,0 +1,5 @@
+function createEventListTemplate() {
+  return ('<ul class="trip-events__list"></ul>');
+}
+
+export {createEventListTemplate};
