+++ src/view/event-list/event-list-view.js
@@ -1,0 +1,20 @@
+import { createElement } from '../../render';
+import { createEventListTemplate } from './event-list-template';
+
+export default class EventListView {
+  getTemplate() {
+    return createEventListTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
