+++ src/view/trip-filters/trip-filters-view.js
@@ -1,0 +1,20 @@
+import { createElement } from '../../render';
+import { createTripFiltersTemplate } from './trip-filters-template';
+
+export default class TripFiltersView {
+  getTemplate() {
+    return createTripFiltersTemplate();
+  }
+
+  getElement() {
+    if (!this.element){
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
