+++ src/view/trip-sort/trip-sort-template.js
@@ -1,5 +1,3 @@
-import { createElement } from '../render';
-
 function createSortTemplate() {
   return (`<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
             <div class="trip-sort__item  trip-sort__item--day">
@@ -29,20 +27,4 @@ function createSortTemplate() {
           </form>`);
 }
 
-export default class TripSortView {
-  getTemplate() {
-    return createSortTemplate();
-  }
-
-  getElement() {
-    if (!this.element){
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
+export {createSortTemplate};
